fix: use uint16/uint32 for Port/BGPPort/LocalASN instead of platform-width int
validateASN compared against the constant 4294967295, which overflows a
32-bit int — a hard compile error on linux/386, linux/arm/v7, and
linux/arm/v5, all of which deploy.yml builds. Switching LocalASN to
uint32 (the actual wire type for a 4-byte BGP ASN) fixes the multi-platform
build and, combined with strconv.ParseUint/encoding/json's built-in range
checks, drops the upper-bound check from validateASN entirely — invalid
values become structurally unrepresentable instead of relying on a
runtime comparison.

Port and BGPPort get the same treatment with uint16, the actual TCP port
type, removing the int32 cast in manager.go's ASN wiring along the way.

+++ internal/web/server.go
@@ -149,9 +149,9 @@ func New(st *settings.Settings, s *store.Store, syncer *feeds.Syncer, bgp BGP) *
 	// show "BGP settings changed — apply now" instead of the change
 	// silently doing nothing until the next process restart.
 	markBGPRestartPending := func() { server.restartPending.Store(true) }
-	st.LocalASN.OnChange(func(int) { markBGPRestartPending() })
+	st.LocalASN.OnChange(func(uint32) { markBGPRestartPending() })
 	st.RouterID.OnChange(func(string) { markBGPRestartPending() })
-	st.BGPPort.OnChange(func(int) { markBGPRestartPending() })
+	st.BGPPort.OnChange(func(uint16) { markBGPRestartPending() })
 	st.LocalAddressV4.OnChange(func(string) { markBGPRestartPending() })
 	st.LocalAddressV6.OnChange(func(string) { markBGPRestartPending() })
